+++ src/lib/core/ace/spinlockguard.hpp
@@ -46,10 +46,10 @@ public:
     }
 
     template <typename T1, typename T2>
-    inline static const auto copyWithLock(spin_lock_t *lock, const T1 &val1, const T2 &val2)
+    inline static auto copyWithLock(spin_lock_t *lock, const T1 &val1, const T2 &val2)
     {
         SpinlockGuard guard(lock);
-        return etl::pair<const T1 &, const T2 &>(val1, val2);
+        return etl::pair<T1, T2>(val1, val2);
     }
 
     operator bool() const